Make gpu_generic_integrator::integrate() respect the dT parameter
This change introduces a subtle difference to the meaning of Tend parameter
(which will in the near future be renamed Tstop) that gets passed to
stopper_t and propagator_t's methods. Tstop is now the "time a at which this
integrator::integrate() call must stop integrating", which does not
necessarily have to be the same as that returned by ensemble::time_end(),
that is "the time beyond this system is not to be integrated".

E.g., if integrating a system from T=[0, 100.5] by repeatedly calling
integrator::integrate() with dT=1, within subsequent integrate() calls Tstop
would be 1, 2, 3, ... 99, 100, 100.5. Note that the last value equals 100.5,
as the time must not exceed ensemble::time_end().