Added the test suite facility (make test)
Added a testing/validation suite facility, implemented as an arbitrary
number of test scripts ending in .test, living in test/.  An example test
script, test/swarm, is given.

Implementation: When the user invokes 'make test', make will call
scripts/run_tests.sh.  This script will remove (if it exists) the
test-outputs directory.  It will then scan (in alphabetical order) all
subdirectories of test/ for executables ending in .test, and execute them
(assuming the executable that was found was test/<testdir>/<exename>.test)
as follows:

1) Set environment variable PREFIX to the full path to base directory of
   swarm source (e.g., swarm sources would reside in $PREFIX/src, etc..)
2) Set envvar TEST to the full path of test directory (usually
   $PREFIX/test/<testdir>)
3) Set envvar RUN to the full path of the test's output directory (usually
   $PREFIX/test-output/<testdir>/<exename)
4) Create the $RUN subdirectory and chdir into it
5) Execute the program. The test will be deemed failed if the program
   returns a nonzero value.
6) All stdin/stderr output will be redirected to $RUN.log

Note that a single <testdir> can hold multiple .test programs (the mechanism
above can handle that, without collisions). If the program needs to store
temporary files, or leave any output for the user to inspect, it should do
so in the $RUN subdirectory.

Typically, test programs will run a certain piece of code and compare it's
output to a "reference copy".  To facilitate easy creation of these
reference copies, if make is run as 'make test-create' it will run exactly
as above with the critical difference being that it will look for and
execute executables ending in .create.  The developer can supply .create
counterparts to their .test scripts that when called will create the
reference copies and store them into $TEST.

A convenient way to do this is to symlink the <exename>.create version to
<exename>.test, test for the name by which the test was called, and if it
was '...create', in stead of comparing to a reference file, store the output
as a reference file.  See test/swarm/swarm-run.test for an example
implementation.